#33 Move LivePreviewListener to folder with admin components
