+++ src/app/(frontend)/layout.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { cn } from 'src/utilities/cn';
 
 import { AdminBar } from '@/components/admin/admin-bar';
-import { LivePreviewListener } from '@/features/previews/components/live-preview-listener';
+import { LivePreviewListener } from '@/components/admin/live-preview-listener';
 import { draftMode } from 'next/headers';
 
 import './globals.css';
